юдна

+++ flaskProject/routes/admin.py
@@ -1,11 +1,14 @@
-from flask import Blueprint, render_template, url_for, redirect, request
-from models import get_db_connection, get_orders, get_order_details, get_products, update_order_status, delete_order
+from flask import Blueprint, render_template, url_for, redirect, request, session, flash
+from models import get_db_connection, get_orders, get_order_details, update_order_status, delete_order
 
 
 admin_bp = Blueprint('admin', __name__)
 
 @admin_bp.route('/admin')
 def admin():
+    if session.get('is_admin') is not True:  # Перевірка, чи є в сесії ключ is_admin
+        flash('У вас немає прав адміна')
+        return redirect(url_for('base'))
     conect = get_db_connection()
     feedback = conect.execute('SELECT * FROM feedback').fetchall()
     users = conect.execute('SELECT * FROM users').fetchall()
@@ -16,6 +19,9 @@ def admin():
 
 @admin_bp.route('/admin/delete_feedback/<int:id>', methods=['POST'])
 def delete_feedback(id):
+    if session.get('is_admin') is not True:  # Перевірка, чи є в сесії ключ is_admin
+        flash('У вас немає прав адміна')
+        return redirect(url_for('base'))
     conn = get_db_connection()
     conn.execute('DELETE FROM feedback WHERE id = ?', (id,))
     conn.commit()
@@ -24,12 +30,18 @@ def delete_feedback(id):
 
 @admin_bp.route('/admin/order/<int:order_id>')
 def details(order_id):
+    if session.get('is_admin') is not True:  # Перевірка, чи є в сесії ключ is_admin
+        flash('У вас немає прав адміна')
+        return redirect(url_for('base'))
     order, items = get_order_details(order_id)
     return render_template('details.html', order=order, items=items)
 
 
 @admin_bp.route('/admin/update_order_status/<int:order_id>', methods=['POST'])
 def update_order(order_id):
+    if session.get('is_admin') is not True:  # Перевірка, чи є в сесії ключ is_admin
+        flash('У вас немає прав адміна')
+        return redirect(url_for('base'))
     status = request.form['status']
     update_order_status(order_id, status)
     return redirect(url_for('admin.admin'))
@@ -37,6 +49,9 @@ def update_order(order_id):
 
 @admin_bp.route('/admin/delete_order/<int:order_id>', methods=['POST'])
 def delete_order_route(order_id):
+    if session.get('is_admin') is not True:  # Перевірка, чи є в сесії ключ is_admin
+        flash('У вас немає прав адміна')
+        return redirect(url_for('base'))
     delete_order(order_id)
     return redirect(url_for('admin.admin'))
